refactor python filters
