+++ examples/custom-gql-apis/src/lib.rs
@@ -4,7 +4,7 @@ use ::raphtory_core::python::packages::base_modules::{
     base_vectors_module,
 };
 use pyo3::prelude::*;
-use raphtory_core::python::types::wrappers::filter_expr::base_filter_module;
+use raphtory::python::filter::base_filter_module;
 use raphtory_graphql::python::pymodule::base_graphql_module;
 
 pub mod python;
