+++ python/python/raphtory/vectors/__init__.pyi
@@ -11,6 +11,7 @@ from __future__ import annotations
 
 from typing import *
 from raphtory import *
+import raphtory.filter as filter
 from raphtory.algorithms import *
 from raphtory.node_state import *
 from raphtory.graphql import *
