+++ raphtory/src/python/filter/node_filter_builders.rs
@@ -1,0 +1,84 @@
+use crate::{
+    db::graph::views::filter::model::{InternalNodeFilterBuilderOps, NodeFilter},
+    python::{
+        filter::filter_expr::PyFilterExpr,
+        types::{iterable::FromIterable, wrappers::prop::DynNodeFilterBuilderOps},
+    },
+};
+use pyo3::{pyclass, pymethods};
+use std::sync::Arc;
+
+/// A builder for constructing node filters
+///
+/// To create a filter builder see [Node][raphtory.filter.Node].
+#[pyclass(frozen, name = "NodeFilterBuilder", module = "raphtory.filter")]
+#[derive(Clone)]
+pub struct PyNodeFilterBuilder(Arc<dyn DynNodeFilterBuilderOps>);
+
+impl<T: InternalNodeFilterBuilderOps + 'static> From<T> for PyNodeFilterBuilder {
+    fn from(value: T) -> Self {
+        PyNodeFilterBuilder(Arc::new(value))
+    }
+}
+
+#[pymethods]
+impl PyNodeFilterBuilder {
+    fn __eq__(&self, value: String) -> PyFilterExpr {
+        self.0.eq(value)
+    }
+
+    fn __ne__(&self, value: String) -> PyFilterExpr {
+        self.0.ne(value)
+    }
+
+    fn is_in(&self, values: FromIterable<String>) -> PyFilterExpr {
+        self.0.is_in(values.into())
+    }
+
+    fn is_not_in(&self, values: FromIterable<String>) -> PyFilterExpr {
+        self.0.is_not_in(values.into())
+    }
+
+    fn contains(&self, value: String) -> PyFilterExpr {
+        self.0.contains(value)
+    }
+
+    fn not_contains(&self, value: String) -> PyFilterExpr {
+        self.0.not_contains(value)
+    }
+
+    fn fuzzy_search(
+        &self,
+        value: String,
+        levenshtein_distance: usize,
+        prefix_match: bool,
+    ) -> PyFilterExpr {
+        self.0
+            .fuzzy_search(value, levenshtein_distance, prefix_match)
+    }
+}
+
+#[derive(Clone)]
+#[pyclass(frozen, name = "Node", module = "raphtory.filter")]
+pub struct PyNodeFilter;
+
+#[pymethods]
+impl PyNodeFilter {
+    /// Filter node by name
+    ///
+    /// Returns:
+    ///     NodeFilterBuilder: A filter builder for filtering by node name
+    #[staticmethod]
+    fn name() -> PyNodeFilterBuilder {
+        PyNodeFilterBuilder(Arc::new(NodeFilter::name()))
+    }
+
+    /// Filter node by type
+    ///
+    /// Returns:
+    ///     NodeFilterBuilder: A filter builder for filtering by node type
+    #[staticmethod]
+    fn node_type() -> PyNodeFilterBuilder {
+        PyNodeFilterBuilder(Arc::new(NodeFilter::node_type()))
+    }
+}
