+++ raphtory/src/python/graph/views/graph_view.rs
@@ -34,11 +34,9 @@ use crate::{
     errors::GraphError,
     prelude::*,
     python::{
+        filter::filter_expr::PyFilterExpr,
         graph::{edge::PyEdge, node::PyNode},
-        types::{
-            repr::{Repr, StructReprBuilder},
-            wrappers::filter_expr::PyFilterExpr,
-        },
+        types::repr::{Repr, StructReprBuilder},
         utils::PyNodeRef,
     },
 };
@@ -355,7 +353,7 @@ impl PyGraphView {
 
     /// Get the edges that match the properties name and value
     /// Arguments:
-    ///     properties_dict (dict[str, Prop]): the properties name and value
+    ///     properties_dict (dict[str, PropValue]): the properties name and value
     /// Returns:
     ///    list[Edge]: the edges that match the properties name and value
     #[pyo3(signature = (properties_dict))]
