+++ raphtory/src/python/mod.rs
@@ -3,6 +3,7 @@ extern crate core;
 #[macro_use]
 pub mod types;
 mod algorithm;
+pub mod filter;
 pub mod graph;
 pub mod packages;
 pub mod utils;
