+++ raphtory/src/python/types/wrappers/mod.rs
@@ -1,5 +1,4 @@
 pub mod document;
-pub mod filter_expr;
 pub mod iterables;
 pub mod iterators;
 pub mod prop;
